chatbot-websocket-used/requirements.txt
openai
fastapi
uvicorn
python-multipart
jinja2
python-dotenv
websockets